Update by separating send_model_params test
